Moved config.json to root
