+++ .gitignore
@@ -5,3 +5,4 @@ jslint.txt
 dist/
 *.mp4
 *.webm
+config.json
