+++ README.md
@@ -16,12 +16,5 @@ npx webpack
 ```
 to compile the `src` directory into bundles. You can also specify the directory that webpack outputs to with `-o output_dir`.
 
-### Developer Server
-The [Webpack-dev-server](https://github.com/webpack/webpack-dev-server) can be enabled by calling the following command.
-```sh
-node_modules/.bin/webpack serve
-```
-This will automatically update when changes are made in `src`.
-
-### Production Server
-In production the server (nginx/apache) can serve the files generated by webpack. Remember to use `--mode="production"` when calling `webpack` to generate the production distribution.
+## Configuration
+Copy the `config.json.example` file to `config.json` and fill in the configuration file. To disable certain tools from being used remove the respective entry from the `tools` key.
