+++ src/js/model/ideasServiceProxy.js
@@ -1,4 +1,4 @@
-import config from '../config.json'
+import config from '../../../config.json'
 import { LogEXSession } from '../logEXSession.js'
 
 /**
